pyproject.toml
[tool.poetry]
name = "agent-studio-backend"
version = "0.1.0"
description = ""
authors = ["g4ngw4r <rakesh@superjackfruit.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
langchain-openai = "^0.1.22"
langchain = "^0.2.14"
supabase = "^2.7.2"
langchain-anthropic = "^0.1.23"
langgraph = "^0.2.4"
langchain-community = "^0.2.12"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
